+++ pression/compressor/compressorZSTD.cpp
@@ -1,5 +1,5 @@
 
-/* Copyright (c) 2012-2016, Stefan Eilemann <eile@eyescale.ch>
+/* Copyright (c) 2016, Stefan Eilemann <eile@eyescale.ch>
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License version 2.1 as published
@@ -15,9 +15,9 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#include "compressorSnappy.h"
+#include "compressorZSTD.h"
 
-#include "snappy/snappy.h"
+#include "zstd/lib/zstd.h"
 
 namespace pression
 {
@@ -25,61 +25,58 @@ namespace plugin
 {
 namespace
 {
-#ifndef __xlC__
 static void _getInfo( EqCompressorInfo* const info )
 {
     info->version = EQ_COMPRESSOR_VERSION;
     info->capabilities = EQ_COMPRESSOR_DATA_1D | EQ_COMPRESSOR_DATA_2D;
     info->quality = 1.f;
-    info->ratio   = .60f;
-    info->speed   = .55f;
-    info->name = EQ_COMPRESSOR_SNAPPY_BYTE;
+    info->ratio   = .47f;
+    info->speed   = .25f;
+    info->name = EQ_COMPRESSOR_ZSTD_BYTE;
     info->tokenType = EQ_COMPRESSOR_DATATYPE_BYTE;
 }
 
 static bool _register()
 {
     Compressor::registerEngine(
-        Compressor::Functions( EQ_COMPRESSOR_SNAPPY_BYTE,
+        Compressor::Functions( EQ_COMPRESSOR_ZSTD_BYTE,
                                _getInfo,
-                               CompressorSnappy::getNewCompressor,
-                               CompressorSnappy::getNewDecompressor,
-                               CompressorSnappy::decompress, 0 ));
+                               CompressorZSTD::getNewCompressor,
+                               CompressorZSTD::getNewDecompressor,
+                               CompressorZSTD::decompress, 0 ));
     return true;
 }
 
 static const bool LB_UNUSED _initialized = _register();
-#endif
 }
 
-void CompressorSnappy::compress( const void* const inData,
-                                 const eq_uint64_t nPixels,
-                                 const bool /*alpha*/ )
+void CompressorZSTD::compress( const void* const inData,
+                               const eq_uint64_t nPixels,
+                               const bool /*alpha*/ )
 {
     _nResults = 1;
     if( _results.size() < _nResults )
         _results.push_back( new pression::plugin::Compressor::Result );
-    size_t size = snappy::MaxCompressedLength( nPixels );
+    size_t size = ZSTD_compressBound( nPixels );
     _results[0]->reserve( size );
 
-    snappy::RawCompress( (const char*)(inData), nPixels,
-                         (char*)( _results[0]->getData( )), &size );
-    assert( size != 0 );
+    size = ZSTD_compress( _results[0]->getData(), size, inData, nPixels, 2 );
     _results[0]->setSize( size );
 }
 
-void CompressorSnappy::decompress( const void* const* inData,
+void CompressorZSTD::decompress( const void* const* inData,
                                    const eq_uint64_t* const inSizes,
                                    const unsigned nInputs,
                                    void* const outData,
-                                   eq_uint64_t* const /*outDims*/,
-                                   const eq_uint64_t, void* const )
+                                   eq_uint64_t* const outDims,
+                                   const eq_uint64_t flags, void* const )
 {
     if( nInputs == 0 )
         return;
 
-    snappy::RawUncompress( (const char*)(inData[0]), inSizes[0],
-                           (char*)(outData) );
+    const eq_uint64_t nPixels = ( flags & EQ_COMPRESSOR_DATA_1D) ?
+                                    outDims[1] : outDims[1] * outDims[3];
+    ZSTD_decompress( outData, nPixels, inData[0], inSizes[0] );
 }
 
 }