+++ doc/Changelog.md
@@ -1,5 +1,9 @@
 # Changelog {#Changelog}
 
+# git master
+
+* [20](https://github.com/Eyescale/Pression/pull/20):
+  Add ZStandard byte compressor
 
 # Version 1.1.1 (8-Mar-2016)
 # Version 1.1 (2-Nov-2015)